display local.approximateDate.issued instead of dc.date.issued
approximateDate is currently either "circa range" (eg. "cca 1920-1949")
or a list of years (each an issue date of a part).
dc.date.issued is either 0000 (unknown) or one of the years listed in
approximate.
This is display only and only in item-view/item-list. The 0000 will be shown in
refbox if you don't use other measures (such as local.refbox.format).

dc.date.issued is not expected to be repeatable; removed the for-each in
item-view.

You need to:
```
bin/dspace registry-loader -metadata .../dspace/config/registries/local-types.xml
```